extract `data-store` from `index`
