+++ package.json
@@ -44,6 +44,7 @@
   },
   "dependencies": {
     "debug": "^2.2.0",
+    "delegates": "^0.1.0",
     "firebase-copy": "0.2.0",
     "jwt-simple": "^0.3.1",
     "lodash": "3.10.1",
